apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project]
name = "tebekaye-api"
version = "0.1.0"
description = "Tebekaye FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.0",
  "python-multipart>=0.0.12",
  "passlib[bcrypt]>=1.7.4",
  "PyJWT>=2.9.0",
  "email-validator>=2.2.0",
  "pypdf>=5.1.0",
  "python-docx>=1.1.2",
  "httpx>=0.27.0",
  "stripe>=11.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
